refactor pipes
