+++ src/codegen/x64/mod.rs
@@ -1,4 +1,4 @@
 pub mod asm;
 pub mod emit;
 pub mod from;
-pub mod pipe;
+pub mod pipes;
