+++ src/codegen/x64/pipes/mod.rs
@@ -1,0 +1,3 @@
+pub mod instruction_fix;
+pub mod pipe;
+pub mod reg_replace;
